requirements.txt
fastapi
uvicorn
pymongo
fashion-clip
pandas
numpy
pillow
IPython
tqdm
passlib
bcrypt